+++ best_flutter_ui_templates/packages/hotel_repository/lib/src/models/hotel.dart
@@ -8,8 +8,8 @@ class Hotel {
   String titleTxt;
   double rating;
   int reviews;
-  double perNight;
-  double discount;
+  int perNight;
+  int discount;
   Location location;
   Hotel({
     required this.hotelId,
